refactor: reorganize core test suite by function
Reorganized test files from phase-based to functional organization
for better maintainability and clarity.

Changes:
- Created functional test files organized by API domain
  - dataset.test.ts (16 tests) - all dataset operations
  - datastore-query.test.ts (14 tests) - all query operations
  - data-dictionary.test.ts (12 tests) - all dictionary operations
  - metastore.test.ts (12 tests) - schema and facets operations

- Updated existing files
  - client.test.ts (18 tests) - focused on core infrastructure only
  - Renamed query-download.test.ts to datastore-download.test.ts

- Removed phase-based test files
  - phase1-openapi-alignment.test.ts
  - phase2-openapi-alignment.test.ts
  - phase3-openapi-alignment.test.ts
  - dataset-crud.test.ts
  - data-dictionary-crud.test.ts

- Added tests for previously untested methods
  - listAllDatasets()
  - getDataDictionary()
  - listDataDictionaries()
  - getDataDictionaryFromUrl()
  - getDatastoreSchema()
  - listSchemas()
  - getOpenApiDocsUrl()

Test count: 140 tests (up from 123)
Coverage: All 42 API methods now have explicit tests
