+++ packages/dkan-client-tools-core/src/__tests__/api/client.test.ts
@@ -1,5 +1,12 @@
 /**
- * Tests for DkanApiClient
+ * Tests for DkanApiClient Core Functionality
+ *
+ * Tests core HTTP client infrastructure:
+ * - Constructor and configuration
+ * - Authentication (token and basic auth)
+ * - Retry logic for network failures
+ * - Error handling and DkanApiError creation
+ * - Utility methods (getBaseUrl, getDefaultOptions, getOpenApiDocsUrl)
  */
 
 import { describe, it, expect, beforeEach, vi, afterEach } from 'vitest'
@@ -10,7 +17,7 @@ import { DkanApiError } from '../../types'
 const mockFetch = vi.fn()
 global.fetch = mockFetch as any
 
-describe('DkanApiClient', () => {
+describe('DkanApiClient - Core Functionality', () => {
   beforeEach(() => {
     mockFetch.mockReset()
   })
@@ -75,44 +82,7 @@ describe('DkanApiClient', () => {
     })
   })
 
-  describe('getDataset', () => {
-    it('should fetch dataset by identifier', async () => {
-      const mockDataset = {
-        identifier: 'test-123',
-        title: 'Test Dataset',
-        description: 'Test description',
-        accessLevel: 'public' as const,
-        modified: '2024-01-01',
-        keyword: ['test'],
-        publisher: { name: 'Test Publisher' },
-        contactPoint: {
-          '@type': 'vcard:Contact',
-          fn: 'Test Contact',
-          hasEmail: 'test@example.com',
-        },
-      }
-
-      mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => mockDataset,
-      })
-
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      const result = await client.getDataset('test-123')
-
-      expect(mockFetch).toHaveBeenCalledWith(
-        'https://example.com/api/1/metastore/schemas/dataset/items/test-123',
-        expect.objectContaining({
-          headers: expect.objectContaining({
-            'Content-Type': 'application/json',
-          }),
-        })
-      )
-      expect(result).toEqual(mockDataset)
-    })
-
+  describe('Authentication', () => {
     it('should include auth header when token provided', async () => {
       mockFetch.mockResolvedValueOnce({
         ok: true,
@@ -163,7 +133,9 @@ describe('DkanApiClient', () => {
         })
       )
     })
+  })
 
+  describe('Error Handling', () => {
     it('should throw DkanApiError on HTTP error', async () => {
       mockFetch.mockResolvedValue({
         ok: false,
@@ -174,182 +146,104 @@ describe('DkanApiClient', () => {
 
       const client = new DkanApiClient({
         baseUrl: 'https://example.com',
-        defaultOptions: { retry: 0 }, // Disable retries for this test
+        defaultOptions: { retry: 0 },
       })
 
       await expect(client.getDataset('nonexistent')).rejects.toThrow(DkanApiError)
       await expect(client.getDataset('nonexistent')).rejects.toThrow('HTTP 404: Not Found')
     })
-  })
 
-  describe('searchDatasets', () => {
-    it('should search datasets with no filters', async () => {
-      const mockResponse = {
-        total: '10',
-        results: {
-          'id-1': { identifier: 'id-1', title: 'Dataset 1' },
-          'id-2': { identifier: 'id-2', title: 'Dataset 2' },
-        },
-      }
-
+    it('should extract error details from JSON response', async () => {
       mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => mockResponse,
+        ok: false,
+        status: 400,
+        statusText: 'Bad Request',
+        text: async () => JSON.stringify({
+          message: 'Validation failed',
+          status: 400,
+          timestamp: '2025-11-13T17:00:00Z',
+          data: {
+            field: 'title',
+            error: 'Required field missing',
+          },
+        }),
       })
 
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      const result = await client.searchDatasets()
-
-      expect(mockFetch).toHaveBeenCalledWith(
-        'https://example.com/api/1/search',
-        expect.any(Object)
-      )
-      expect(result.total).toBe(10)
-      expect(result.results).toHaveLength(2)
-      expect(result.results[0]).toEqual({ identifier: 'id-1', title: 'Dataset 1' })
-    })
-
-    it('should search datasets with keyword filter', async () => {
-      mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => ({ total: 5, results: {} }),
+      const client = new DkanApiClient({
+        baseUrl: 'https://example.com',
+        defaultOptions: { retry: 0 },
       })
 
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      await client.searchDatasets({ keyword: 'health' })
-
-      expect(mockFetch).toHaveBeenCalledWith(
-        'https://example.com/api/1/search?keyword=health',
-        expect.any(Object)
-      )
-    })
-
-    it('should search datasets with multiple filters', async () => {
-      mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => ({ total: 3, results: {} }),
-      })
-
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      await client.searchDatasets({
-        keyword: 'health',
-        theme: 'Health',
-        page: 1,
-        'page-size': 20,
-        sort: 'modified',
-        'sort-order': 'desc',
-      })
-
-      expect(mockFetch).toHaveBeenCalledWith(
-        'https://example.com/api/1/search?keyword=health&theme=Health&sort=modified&sort-order=desc&page=1&page-size=20',
-        expect.any(Object)
-      )
-    })
-
-    it('should transform results object to array', async () => {
-      const mockResponse = {
-        total: 2,
-        results: {
-          'abc': { identifier: 'abc', title: 'First' },
-          'def': { identifier: 'def', title: 'Second' },
-        },
-      }
-
-      mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => mockResponse,
-      })
-
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      const result = await client.searchDatasets()
-
-      expect(Array.isArray(result.results)).toBe(true)
-      expect(result.results).toHaveLength(2)
-    })
-
-    it('should parse total as number when string', async () => {
-      mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => ({ total: '42', results: {} }),
-      })
-
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      const result = await client.searchDatasets()
-
-      expect(result.total).toBe(42)
-      expect(typeof result.total).toBe('number')
-    })
-  })
-
-  describe('queryDatastore', () => {
-    it('should query datastore with POST request', async () => {
-      const mockResponse = {
-        results: [{ field1: 'value1' }],
-        count: 1,
-      }
-
-      mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => mockResponse,
-      })
-
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      const result = await client.queryDatastore('dataset-123', 0, {
-        limit: 10,
-        offset: 0,
-      })
-
-      expect(mockFetch).toHaveBeenCalledWith(
-        'https://example.com/api/1/datastore/query/dataset-123/0',
-        expect.objectContaining({
-          method: 'POST',
-          body: JSON.stringify({ limit: 10, offset: 0 }),
+      try {
+        await client.getDataset('invalid-id')
+        expect.fail('Should have thrown error')
+      } catch (error) {
+        expect(error).toBeInstanceOf(DkanApiError)
+        const apiError = error as DkanApiError
+        expect(apiError.message).toBe('Validation failed')
+        expect(apiError.statusCode).toBe(400)
+        expect(apiError.timestamp).toBe('2025-11-13T17:00:00Z')
+        expect(apiError.data).toEqual({
+          field: 'title',
+          error: 'Required field missing',
         })
-      )
-      expect(result).toEqual(mockResponse)
+      }
     })
 
-    it('should query datastore with complex options', async () => {
-      const options = {
-        conditions: [{ property: 'status', value: 'active' }],
-        properties: ['id', 'name'],
-        sorts: [{ property: 'name', order: 'asc' as const }],
-        limit: 50,
-      }
-
+    it('should handle error response without timestamp', async () => {
       mockFetch.mockResolvedValueOnce({
-        ok: true,
-        status: 200,
-        statusText: 'OK',
-        json: async () => ({ results: [], count: 0 }),
+        ok: false,
+        status: 404,
+        statusText: 'Not Found',
+        text: async () => JSON.stringify({
+          message: 'Dataset not found',
+        }),
       })
 
-      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
-      await client.queryDatastore('dataset-123', 0, options)
+      const client = new DkanApiClient({
+        baseUrl: 'https://example.com',
+        defaultOptions: { retry: 0 },
+      })
 
-      expect(mockFetch).toHaveBeenCalledWith(
-        expect.any(String),
-        expect.objectContaining({
-          method: 'POST',
-          body: JSON.stringify(options),
-        })
-      )
+      try {
+        await client.getDataset('missing-id')
+        expect.fail('Should have thrown error')
+      } catch (error) {
+        expect(error).toBeInstanceOf(DkanApiError)
+        const apiError = error as DkanApiError
+        expect(apiError.message).toBe('Dataset not found')
+        expect(apiError.timestamp).toBeUndefined()
+        expect(apiError.data).toBeUndefined()
+      }
+    })
+
+    it('should handle non-JSON error responses', async () => {
+      mockFetch.mockResolvedValueOnce({
+        ok: false,
+        status: 500,
+        statusText: 'Internal Server Error',
+        text: async () => 'Server error occurred',
+      })
+
+      const client = new DkanApiClient({
+        baseUrl: 'https://example.com',
+        defaultOptions: { retry: 0 },
+      })
+
+      try {
+        await client.searchDatasets({})
+        expect.fail('Should have thrown error')
+      } catch (error) {
+        expect(error).toBeInstanceOf(DkanApiError)
+        const apiError = error as DkanApiError
+        expect(apiError.message).toContain('500')
+        expect(apiError.statusCode).toBe(500)
+        expect(apiError.timestamp).toBeUndefined()
+      }
     })
   })
 
-  describe('Retry logic', () => {
+  describe('Retry Logic', () => {
     it('should retry on network failure', async () => {
       vi.useFakeTimers()
 
@@ -390,13 +284,13 @@ describe('DkanApiClient', () => {
 
       const client = new DkanApiClient({
         baseUrl: 'https://example.com',
-        defaultOptions: { retry: 2, retryDelay: 1 }, // Use minimal delay with real timers
+        defaultOptions: { retry: 2, retryDelay: 1 },
       })
 
       await expect(client.getDataset('test')).rejects.toThrow(DkanApiError)
       expect(mockFetch).toHaveBeenCalledTimes(3) // Initial + 2 retries
 
-      mockFetch.mockReset() // Clean up for next test
+      mockFetch.mockReset()
     })
 
     it('should not retry on HTTP errors (only network errors)', async () => {
@@ -410,13 +304,45 @@ describe('DkanApiClient', () => {
 
       const client = new DkanApiClient({
         baseUrl: 'https://example.com',
-        defaultOptions: { retry: 0 }, // No retries to avoid complexity
+        defaultOptions: { retry: 0 },
       })
 
       await expect(client.getDataset('test')).rejects.toThrow(DkanApiError)
 
       // Should only call fetch once (no retries for HTTP errors)
       expect(mockFetch).toHaveBeenCalledTimes(1)
+    })
+  })
+
+  describe('Utility Methods', () => {
+    it('should return base URL', () => {
+      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
+      expect(client.getBaseUrl()).toBe('https://example.com')
+    })
+
+    it('should return default options', () => {
+      const client = new DkanApiClient({
+        baseUrl: 'https://example.com',
+        defaultOptions: {
+          retry: 5,
+          retryDelay: 2000,
+        },
+      })
+      const options = client.getDefaultOptions()
+      expect(options.retry).toBe(5)
+      expect(options.retryDelay).toBe(2000)
+    })
+
+    it('should return OpenAPI docs URL', () => {
+      const client = new DkanApiClient({ baseUrl: 'https://example.com' })
+      const docsUrl = client.getOpenApiDocsUrl()
+      expect(docsUrl).toBe('https://example.com/api/1')
+    })
+
+    it('should handle baseUrl with trailing slash in OpenAPI docs URL', () => {
+      const client = new DkanApiClient({ baseUrl: 'https://example.com/' })
+      const docsUrl = client.getOpenApiDocsUrl()
+      expect(docsUrl).toBe('https://example.com/api/1')
     })
   })
 })
